rock_paper_scissors/pyproject.toml
[project]
name = "rock-paper-scissors"
version = "0.1.0"
description = "A Python game where the player competes against the computer by selecting rock, paper, or scissors, and the program decides the winner based on their choices."
authors = [
    {name = "ktsonyovski",email = "kaloyan.tsonyovski@festo.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pylint (>=3.3.7,<4.0.0)",
    "pytest-cov (>=6.2.1,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
